Header style/structure complete with typewriter
Created the website header with a typewriter effect for the title.

I allowed printing to console with .tslint so I can see my logging.

I tried to update a property using the KeyOf function, but it failed.
I can access the property, but I cannot update its value.
Not sure why... I need to move on, but an interesting problem to
figure out in the future.