BaseComponent and home structure
I made a base component. The plan is to inheret from the base component
for each of the component sections. Hopefully this will provide me with
a library of components that I can customise quickly if I ever need to
rebuild another website in a similar structure.

This commit should work as a template project for further websites.
So long as they need a simple 1 page layout. This can at least be a
version 1 template for myself.

I need to work on making is simpler to customise as some of this is very
much only appropriate for my personal website.

I have just made a base component. Still need to prove that this can
be inherited from in a smart way.